Move files in frontend to make it more presentable
