+++ frontend/src/App.js
@@ -5,16 +5,15 @@ import {
   Switch,
   Route
 } from 'react-router-dom'
-import homepage from './component/homepage'
-import room from './component/room'
+import homepage from './component/homepage/homepage'
+import room from './component/room/room'
 
 import SignIn from './component/Auth/signin';
 import SetToken from './component/Auth/setAuth';
-import SignOut from './component/Auth/logout'
 import {useAuth} from './context/AuthContext'
 import {CookiesProvider} from "react-cookie"
-import PageNotFound from './component/404Page'
-import unathorised from './component/unAuthorised'
+import PageNotFound from './error/404Page'
+import unathorised from './error/unAuthorised'
 import Header from './component/navbar';
 function App() {
 
@@ -30,7 +29,7 @@ function App() {
               <Route exact path={'/room/:difficulty/:id'} component={room}></Route>
               <Route exact path={'/login'} component={SignIn}></Route>
               <Route path={'/setAuth'} login={login} loggedIn={loggedIn} logout={logout} user={user} token={token} component={SetToken}></Route>
-              <Route path={'/logout'} component={SignOut}></Route>
+
               <Route exact path={'/unauthorised'} component={unathorised}></Route>
               <Route component={PageNotFound} />
             </Switch>
