+++ frontend/src/component/room/partnerInfo.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useAuth } from  "../context/AuthContext";
+import { useAuth } from  "../../context/AuthContext";
 
 const Cookies = require("js-cookie");
